Add fork to command palette
