+++ crates/pcb-zen/src/lib.rs
@@ -6,6 +6,7 @@ mod auto_deps;
 pub mod cache_index;
 pub mod canonical;
 pub mod diagnostics;
+pub mod fork;
 pub mod git;
 pub mod load;
 pub mod lsp;
